requirements.txt
absl-py==1.4.0
altair==4.2.2
appdirs==1.4.4
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attr==0.3.1
attrs==23.1.0
azure-core==1.26.4
azure-storage-blob==12.16.0
backcall==0.2.0
bleach==5.0.1
blinker==1.6.2
boto==2.49.0
boto3==1.26.138
botocore==1.29.138
boxing==0.1.4
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
cycler==0.10.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.16
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
entrypoints==0.4
executing==1.2.0
expiringdict==1.2.2
filelock==3.12.0
fire==0.5.0
Flask==2.3.2
flatbuffers==23.5.9
fonttools==4.39.4
fsspec==2023.5.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.10.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.3
grpcio==1.54.2
grpcio-status==1.48.2
h5py==3.8.0
htmlmin==0.1.12
huggingface-hub==0.14.1
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
inflection==0.5.1
ipykernel==6.23.1
ipython==8.13.2
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==3.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.12.0
kiwisolver==1.4.4
label-studio-converter==0.0.51
label-studio-tools==0.0.2
launchdarkly-server-sdk==7.5.0
libclang==16.0.0
lockfile==0.12.2
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.6.7
numpy==1.21.6
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==23.1
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
pytz==2019.3
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
redis==4.5.5
regex==2023.5.5
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.3.5
roboflow==1.0.9
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.6.1
sahi==0.11.14
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semver==2.13.0
sentry-sdk==1.23.1
shapely==2.0.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
stack-data==0.6.2
streamlit==1.22.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
ua-parser==0.16.1
ujson==5.7.0
ultralytics==8.0.107
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.4
wget==3.2
wrapt==1.14.1
xmljson==0.2.0
yolov5==7.0.12
zipp==3.15.0
